requirements.txt
absl-py==0.8.1
alabaster==0.7.12
argmagic==2017.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
blis==0.2.4
boto3==1.9.230
botocore==1.12.230
Bottleneck==1.2.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
conllu==1.3.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.10
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
fastai==1.0.59
fastprogress==0.1.21
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
future==0.18.2
gast==0.2.2
gevent==1.4.0
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grammarbot==0.2.0
graphs==0.1.3
greenlet==0.4.15
grpcio==1.25.0
h5py==2.10.0
hdbscan==0.8.23
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.23
insanity==2017.1
ipdb==0.12.2
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
llvmlite==0.30.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
murmurhash==1.0.2
natsort==6.0.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nlpnet==1.2.4
nltk==3.4.5
notebook==6.0.1
numba==0.46.0
numexpr==2.7.0
numpy==1.17.3
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.1.0
overrides==1.9
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
pbr==5.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.5
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pybind11==2.4.3
Pygments==2.4.2
pyparsing==2.4.2
pypng==0.0.19
pyrsistent==0.15.5
pytest==5.1.2
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.1.0
qtconsole==4.5.5
rdflib==4.2.2
regex==2019.8.19
reldata==2017.1
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.10.6
rsa==4.0
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.13.0
sk-video==1.1.10
sklearn==0.0
snowballstemmer==1.9.1
sonnet==0.1.6
soupsieve==1.9.5
spacy==2.1.8
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.1.0
stanfordnlp==0.2.0
staticinit==2017.1
tensorboard==2.0.1
tensorboardX==1.8
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.3
thinc==7.0.8
torch==1.1.0
torchvision==0.2.2.post3
tornado==6.0.3
tqdm==4.36.0
traitlets==4.3.3
umap-learn==0.3.10
Unidecode==1.1.1
urllib3==1.25.6
virtualenv==16.7.7
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.11.2
zipp==0.6.0
